refactor: :art: more refactors and a readme
moves TotalSummary Card to its own folder and adds a readme
