+++ frontend/src/components/BudgetLineItems/BudgetLinesTable/BudgetLinesTable.jsx
@@ -1,5 +1,5 @@
 import PropTypes from "prop-types";
-import TotalSummaryCard from "./TotalSummaryCard";
+import TotalSummaryCard from "../TotalSummaryCard/TotalSummaryCard";
 import Table from "../../UI/Table";
 import "./BudgetLinesTable.scss";
 import { BUDGET_LINE_TABLE_HEADERS } from "../../../constants";
