+++ frontend/src/components/BudgetLineItems/SummaryCardsSection/SummaryCardsSection.js
@@ -1,0 +1,5 @@
+const SummaryCardsSection = () => {
+    return <p>SummaryCardsSection</p>;
+};
+
+export default SummaryCardsSection;
