+++ frontend/src/components/BudgetLineItems/TotalSummaryCard/index.js
@@ -1,0 +1,1 @@
+export { default } from "./TotalSummaryCard";
