+++ frontend/src/components/BudgetLineItems/readme.md
@@ -1,0 +1,4 @@
+# Budget Line Items
+
+This directory will contain all components related to budget lines.
+Some examples are the `BudgetLinesTable` for Agreements details budget lines tab and the `AllBudgetLinesTable` and `AllBLIRow` seen on the Budget Lines page.
